+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 * Adjust the FOLIO record type in the field mapping profile (UIDATIMP-1027)
 * Adjust the FOLIO record type in the action profile (UIDATIMP-1028)
 * Cover `<ViewContainer>` component with tests (UIDATIMP-731)
+* Cover `<ActionProfiles>` settings with tests (UIDATIMP-971)
+* Use correct `css-loader` syntax (UIDATIMP-1037)
+* Cover `getCRUDActions` with tests (UIDATIMP-1032)
+* Settings : Updates to default job profiles (UIDATIMP-1024)
+* Setting: Updates to Action profiles (UIDATIMP-1026)
+* Adjust the FOLIO record type in the match profile (UIDATIMP-1029)
+* Settings : Updates to default SRS Holdings job profile (UIDATIMP-1023)
 
 ## [5.0.1](https://github.com/folio-org/ui-data-import/tree/v5.0.1) (2021-10-19)